refactor: Community 구조 개선
