+++ src/main/java/com/team05/linkup/common/config/SecurityConfig.java
@@ -53,7 +53,7 @@ public class SecurityConfig {
 
                 .authorizeHttpRequests(authorizeRequests -> authorizeRequests
                         .requestMatchers("/").permitAll()
-                        .requestMatchers("/index.html").permitAll()
+                        .requestMatchers("/static/index.html").permitAll()
                         .requestMatchers("/swagger-ui/**", "/v3/api-docs/**").permitAll()
                         .requestMatchers("/oauth2/authorization/google/**").permitAll()
                         .requestMatchers("/oauth2/authorization/kakao/**").permitAll()
