+++ src/main/java/com/team05/linkup/domain/community/domain/Community.java
@@ -1,5 +1,5 @@
-package com.team05.linkup.domain;
-
+package com.team05.linkup.domain.community.domain;
+import com.team05.linkup.domain.user.domain.User;
 import com.team05.linkup.domain.baseEntity.BaseEntity;
 import com.team05.linkup.domain.enums.CommunityCategory;
 import jakarta.persistence.*;
