+++ src/main/java/com/team05/linkup/domain/community/infrastructure/CommunityRepository.java
@@ -1,6 +1,6 @@
-package com.team05.linkup.community.core.repository;
+package com.team05.linkup.domain.community.infrastructure;
 
-import com.team05.linkup.domain.Community;
+import com.team05.linkup.domain.community.domain.Community;
 import com.team05.linkup.domain.enums.CommunityCategory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,9 +16,9 @@ public interface CommunityRepository extends JpaRepository<Community, String> {
 
     Page<Community> findAllByOrderByViewCountDesc(Pageable pageable);
 
-    @Query("SELECT c FROM Community c LEFT JOIN Like l ON c.id = l.community.id " +
-            "GROUP BY c.id ORDER BY COUNT(l.id) DESC")
-    Page<Community> findAllOrderByLikeCountDesc(Pageable pageable);
+//    @Query("SELECT c FROM Community c LEFT JOIN Like l ON c.id = l.community.id " +
+//            "GROUP BY c.id ORDER BY COUNT(l.id) DESC")
+//    Page<Community> findAllOrderByLikeCountDesc(Pageable pageable);
 
     @Query("SELECT c FROM Community c WHERE c.title LIKE %:keyword% OR c.content LIKE %:keyword%")
     Page<Community> findByTitleContainingOrContentContaining(@Param("keyword") String keyword, Pageable pageable);
